+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Latest
 
+* [minor] - Added warning for `quote-props` to become error in next major release
+
 ## 1.1.0
 
 * Added Changelog
 * [minor] - Dropping `space-infix-ops` rule for now
 * [minor] - Removed unused `promise/no-native`
 * [minor] - Removed `no-else-return`
+* [minor] - Added warnings for rules added by `eslint:recommend` to become errors in next major release
 * Now inherits from `eslint:recommend`
 
 ### 1.0.2